chore:Split files by page component
