+++ src/components/pages/main.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import Header from '../organisms/header';
+import Content from '../organisms/content';
+
+const Main: React.FC = () => {
+  return(
+    <div>
+      <Header />
+      <Content />
+    </div>
+  )
+}
+
+export default Main;
